Add basic statistics endpoints
This commit adds two API endpoints: one for the total number of
conversions and another for other basic statistics in a single
endpoint. The first endpoint, for the number of conversions, is also
used on the frontend. For these purposes, the commit includes the
implementation and schema for the SQLite database, along with the
necessary operations.