Add budget mode
This commit adds the base version of the budget mode. In this mode, the
user is given a randomly generated amount of money for each currency
and can only convert the funds available in their account. An option to
convert all the money from a given currency is also available.